requirements.txt
instaloader==4.10.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3